Rework the navigation components routes. Add Contractor Finish Task List
